+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- LinkedIn icon button functionality; link LinkedIn button to LinkedIn homepage
+- Github icon button functionality; link Github button to repo page
+- LinkedIn and Github icon button animations
+
+### Changed
+
+- Portfolio button to match LinkedIn and Github button style
+
 ## [0.0.1] - 2024-07-01
 
 ### Added
@@ -32,4 +44,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Color previews are provided above each color input with an included hex color corresponding to the selected background or UI color.
 - App component passes a callback function as a property to get updated JavaScript Object data from child components Navbar and SettingMenu and pass updated time format and position values to child component Clock.
 
+[unreleased]: https://github.com/TimochiL/cuztomizable-clock-app/compare/v0.0.1...HEAD
 [0.0.1]: https://github.com/TimochiL/cuztomizable-clock-app/releases/tag/v0.0.1